Refactor hotel page components and remove unused files
- Deleted unused components: GuestReviews, HotelGallery, HotelHeader, HotelInfoPrices, RoomSelection, and HotelNavigation.
- Updated imports in HotelPage to reflect new component structure.
- Improved organization of hotel page components for better maintainability.
