+++ client/src/components/hotelPage/gallery/HotelGallery.tsx
