+++ client/src/components/hotelPage/header/HotelHeader.tsx
