+++ client/src/components/hotelPage/info/EmptyState.tsx
@@ -1,0 +1,29 @@
+export default function EmptyState() {
+  return (
+    <div className="flex flex-1 items-center justify-center py-12">
+      <div className="text-center">
+        <div className="mx-auto mb-4 h-16 w-16 rounded-full bg-gray-100 flex items-center justify-center">
+          <svg
+            className="h-8 w-8 text-gray-400"
+            fill="none"
+            stroke="currentColor"
+            viewBox="0 0 24 24"
+          >
+            <path
+              strokeLinecap="round"
+              strokeLinejoin="round"
+              strokeWidth={2}
+              d="M19 21V5a2 2 0 00-2-2H7a2 2 0 00-2 2v16m14 0h2m-2 0h-5m-9 0H3m2 0h5M9 7h1m-1 4h1m4-4h1m-1 4h1m-5 10v-5a1 1 0 011-1h2a1 1 0 011 1v5m-4 0h4"
+            />
+          </svg>
+        </div>
+        <h3 className="text-lg font-semibold text-gray-900 mb-2">
+          Нет свободных комнат
+        </h3>
+        <p className="text-gray-600">
+          К сожалению, в данный момент нет доступных комнат для выбранных дат.
+        </p>
+      </div>
+    </div>
+  );
+}
