+++ client/src/components/hotelPage/navigation/HotelNavigation.tsx
