+++ client/src/components/hotelPage/rooms/RoomSelection.tsx
@@ -1,4 +1,4 @@
-import { useState, useEffect, useRef } from "react";
+import React, { useState, useEffect, useRef } from "react";
 import { Button } from "@/components/ui/button";
 import { MdBed, MdBathtub, MdSquare } from "react-icons/md";
